fix: renamed widget to WidgetFooter across repo
